requirements.txt
streamlit==1.56.0
pandas==3.0.2
numpy==2.4.4
requests==2.33.1
plotly==6.6.0
matplotlib==3.10.8
rosetta-soil==0.3.2
streamlit-js-eval==1.0.0
